split ElementInspector into multiple files
